requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
apimatic-core==0.2.0
apimatic-core-interfaces==0.1.4
apimatic-requests-client-adapter==0.1.6
asana==5.0.3
asgiref==3.7.2
astroid==3.0.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
CacheControl==0.12.14
cachetools==5.3.2
certifi==2023.11.17
cffi==1.17.0
charset-normalizer==3.1.0
click==8.1.7
commonmark==0.9.1
contourpy==1.3.0
cryptography==42.0.8
cycler==0.12.1
deprecation==2.1.0
detect-secrets==1.5.0
dill==0.3.7
discord.py==2.3.2
distro==1.9.0
dokuwiki==1.3.3
enum34==1.1.10
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
flask-sock==0.7.0
fonttools==4.54.1
frozenlist==1.4.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
gunicorn==21.2.0
guppy3==3.1.4.post1
h11==0.14.0
holidays==0.45
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.6
importlib-metadata==6.8.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.5.0
jsonpickle==3.0.2
jsonpointer==2.4
kiwisolver==1.4.7
mail-parser==3.15.0
Markdown==3.6
MarkupSafe==2.1.3
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
matplotlib==3.9.2
mccabe==0.7.0
memory-profiler==0.61.0
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
numpy==2.1.1
oauthlib==3.2.2
openai==1.51.0
packaging==23.2
pillow==10.4.0
platformdirs==3.0.0
pluggy==1.5.0
protobuf==4.25.1
psutil==6.0.0
PuLP==2.8.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.6.0
pylint==3.0.3
pyparsing==3.1.1
pytest==8.3.3
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
squareup==38.1.0.20240919
tomli==2.0.1
tomlkit==0.12.3
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
